close #15 条件検索をコンポーネント化する
